pyproject.toml
[tool.poetry]
name = "high-order-generative-music"
version = "0.1.0"
description = ""
authors = ["John <john.loverich@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.12.0"
torchaudio = "^0.12.0"
mlflow = "^1.27.0"
hydra-core = "^1.2.0"
matplotlib = "^3.5.2"
torch-optimizer = "^0.3.0"
pytorch-lightning = "^1.6.4"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
